+++ library/prophecy/ProphecyFacts.v
@@ -1,5 +1,5 @@
 From CRIS.common Require Export CRIS.
-From CRIS.simulations.filter Require Export CallFilter.
+From CRIS.filter Require Export CallFilter.
 From CRIS.prophecy Require Export ProphecyIAproof.
 From stdpp Require Import base list.