+++ CHANGELOG.md
@@ -5,6 +5,10 @@ release.
 
 ## Unreleased
 
+- rename: `closed_adequacy` -> `ISim_closed_adequacy`
+- add `gsim_closed_adequacy`, `lsim_closed_adequacy`
+- move `theories/simulations/filter` to `theories/filter`
+
 ## 2026-07-22
 
 ### Added